Cleaned up Implicit Grant Workflow
All requests for the web app go to a login page, which runs a script and redirects to PureCloud's login if there is no access token in the hash. If there is an access token in the hash, it is sent to the server, where it will be checked against the PureCloud API to see if it is valid.

If the token is valid, the user is granted access to the application. If it is not valid, they are redirected to the login page.