Added caching of session data with Redis
Access Tokens, X-OrgBook-Auth-Keys (for mobile), and basic data about the user is now stored upon successful login.

Middleware was added to check that the tokens/keys exist in the Redis database before allowing operations to be performed when interacting routes within /events. The middleware retrieves the information stored about the users for use in the main operation of the route.