Changed Mongo schema and added service layer
Users are now CheckIns and is no longer a subdocument of an Event. It now has its own collection.

A service layer was added and placed in-between the controllers and models to handle business logic.

A route was added to allow attendees to be added to private events. A route was also added to allow a user to be checked in.

A route was also added to allow check-ins. If an event is public, a check-in is inserted into the Mongo collection. If the event is private, the system checks to see if a record of the check-in has been made before updating the check-in's status.