Modified Tests for EventDao
Added newer tests for EventDao and changed existing tests to fit the changes made to the DAO in the last commit.

Also added a method to the dao that allows the removal of a check-in by specifying the personID and eventID. Checkins can also be removed in bulk by specifying just the eventID.

SessionStoreService's storeSessionData method was changed to use a callback (previously it just ran async, meaning that if something went wrong and tokens/auth keys could not be stored, there would be no way to tell). The PureCloudAPIService's login method was changed to use the callback as a result.