Added implemented basic routing for dashboard
Routing was implemented using the react-router package. The hash history object is placed in its own module and passed to the Router component so that other components/objects can have the ability to route to other pages. This also allows components to be implemented using ES6 classes because it removes the need for mixins.