Managed Events are now cached
Multiple requests to Mongo Database only happen on the first request. The eventID is cached so that if a user is updating an event or checking in a user, a request does not need to be made to mongo to check if the user has management access.

Also fixed problem with pagination.