Made modifications to how check-ins are handled.
Dates are now sent back as the time in milliseconds.

When checking a user into a public event, it will initially attempt to insert the check in. If that fails, it will try to update. This is only done to allow managers to check in.

Also added routes to API Docs. Changed "getAttendees" to "getEventCheckIns".

Also changed README to include a link to the Swagger API Docs.