Revised Documentation to make it cleaner
Also changed how QRCodes are retrieved. When retrieving a QRCode, the access_token is sent through the request's query instead of the header. This is to make it easier to generate codes and place images in an img tag.

Tests were also added to ensure that the PureCloudAPIService functions correctly.