Separated Frontend from backend
Frontend code has been taken out and moved to a separate repo.

An access token is now verified against Purecloud and (if successfully authenticated) a session is stored in redis after the first request to the server. All requests that happen afterwards just checks redis to see if token exists.