Added a way to send invitations to users
Invites are stored in a separate table to allow for the storage of invitation IDs. It (hopefully) should make for easier queries and will help prevent too much information from being exposed in a URL (such as personID and eventID). It may be easier to add more features to the emails in the future too.

A separate data access object and service was created to handle anything to do with invitations.

Added email template that is loaded into memory when the application loads. Data is rendered using mustache.

Emails are able to be sent using nodemailer. Currently, the system is not set up to send emails upon adding an attendee.

Some naming changes were made to the schema and to some of the data that is stored in the cache.

Added orgName to events so that it can be used when sending emails (it was originally stored in a user's session data anyway).