Implemented a way to respond to invites
Added an invite response page and 404 page. When the event's start date has been passed, the response to the invite will be rejected and the user will be sent to the 404 Not Found page.

Also made it so that invites related to an event are asynchronously removed from the DB when an event is removed.