Added Check-in search and query for dates between ranges
Checkins can be searched now by field and by inviteStatus (if the event was private). Also added field filtering for event searching too.

A route was added for getting all events that a user has access to between two dates.

Modifications were made to the upcoming filter to allow ongoing events to be returned as well.