Added NotificationService class
Service is used for sending, storing, and retrieving notifications for a user.
New notifications to show to the user are determined by the user's lastSeen date.

Changes were also made to the error handling. There is now a CustomError that is created by
all functions that send back a more detailed error.