Added option for choosing whether to store a notification or not
Made slight changes to the names of some of the notification constants
to better specify the action they convey.

Also changed how the date filter works for db calls used by the /events/getEvents
route. To allow events that start before the 'after' field and end after the 'before'
field to be returned. This was mainly done for the calendar on the frontend.